app-crm/requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
PyYAML>=6.0
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.6
httpx>=0.25
